Filter fixes, tests, and related updates

+++ front-end/src/pages/Evaluator/utils/evaluatorSearchSchema.ts
@@ -1,13 +1,13 @@
+import { ITEMS_PER_PAGE } from '@src/constants/settings'
 import { fallback } from '@tanstack/router-zod-adapter'
 import { z } from 'zod'
-import { ITEMS_PER_PAGE } from '../../../constants/settings'
 import BooleanStringValidator from './booleanValidator'
 import { listValueValidator } from './listValidator'
 import minMaxValidator from './minMaxValidator'
 
 export const evaluatorSchema = z.object({
   view: fallback(z.enum(['all', 'sample']), 'sample'),
-  page: fallback(z.number().gt(0), 1),
+  page: fallback(z.number().gt(0).int(), 1),
   page_size: fallback(z.number().gt(0), ITEMS_PER_PAGE),
   amt_past_due_min: minMaxValidator,
   amt_past_due_max: minMaxValidator,
